Let only admin to edit their channels

+++ src/chat/endpoints.rs
@@ -2,12 +2,12 @@ use rocket::serde::json::Json;
 use rocket_db_pools::Connection;
 use crate::database::{Db, AuthDatabase};
 use crate::models;
-use crate::models::{Channel, ChannelError, ChannelId};
+use crate::models::{Channel, ChannelError, ChannelId, User};
 use crate::schema::messages::channel_id;
 use crate::database::channels::{Database, DataInsertionError, DataRemovalError, DataRetrievalError, DataSetError};
 
 #[get("/channels/<id>")]
-pub async fn get_channel_by_id(id: models::ChannelId, mut db: Connection<Db>) -> Result<Channel, ChannelError> {
+pub async fn get_channel_by_id(id: models::ChannelId, user: User, mut db: Connection<Db>) -> Result<Channel, ChannelError> {
     db.get_channel(id.into())
         .await
         .map_err(|e| match e {
@@ -17,7 +17,19 @@ pub async fn get_channel_by_id(id: models::ChannelId, mut db: Connection<Db>) ->
 }
 
 #[patch("/channels/<id>", format = "json", data = "<patch>")]
-pub async fn patch_channel_by_id(id: models::ChannelId, patch: models::ChannelPatch, mut db: Connection<Db>) -> Result<Channel, ChannelError> {
+pub async fn patch_channel_by_id(id: models::ChannelId, user: User, patch: models::ChannelPatch, mut db: Connection<Db>) -> Result<Channel, ChannelError> {
+    //let channel = get_channel_by_id(id, user, db).await?;
+    let channel = db.get_channel(id.into())
+        .await
+        .map_err(|e| match e {
+            DataRetrievalError::NotFound => ChannelError::NotFound,
+            DataRetrievalError::InternalError => ChannelError::InternalServerError,
+        })?;
+
+    if channel.admin_id != user.id {
+        return Err(ChannelError::Forbidden);
+    }
+
     db.patch_channel(id.into(), patch)
         .await
         .map_err(|e| match e {
@@ -26,7 +38,18 @@ pub async fn patch_channel_by_id(id: models::ChannelId, patch: models::ChannelPa
 }
 
 #[delete("/channels/<id>")]
-pub async fn remove_channel_by_id(id: models::ChannelId, mut db: Connection<Db>) -> Result<Channel, ChannelError> {
+pub async fn remove_channel_by_id(id: models::ChannelId, user: User, mut db: Connection<Db>) -> Result<Channel, ChannelError> {
+    let channel = db.get_channel(id.into())
+        .await
+        .map_err(|e| match e {
+            DataRetrievalError::NotFound => ChannelError::NotFound,
+            DataRetrievalError::InternalError => ChannelError::InternalServerError,
+        })?;
+
+    if channel.admin_id != user.id {
+        return Err(ChannelError::Forbidden);
+    }
+
     db.remove_session(id.into())
         .await
         .map_err(|e| match e {
@@ -35,7 +58,8 @@ pub async fn remove_channel_by_id(id: models::ChannelId, mut db: Connection<Db>)
 }
 
 #[post("/channels", format = "json", data = "<channel>")]
-pub async fn create_channel(channel: models::ChannelInsert, mut db: Connection<Db>) -> Result<Channel, ChannelError> {
+pub async fn create_channel(mut channel: models::ChannelInsert, user: User, mut db: Connection<Db>) -> Result<Channel, ChannelError> {
+    channel.admin_id = user.id;
     db.insert_channel(channel)
         .await
         .map_err(|e| match e {
